[ELY-2005] Adjust Roles to Group conversion so it only happens if the interface 'java.security.acl.Group' is actually available.
